requirements.txt
pandas
numpy
scikit-learn==1.7.1
xgboost==3.0.5
joblib

fastapi
uvicorn
pydantic
python-multipart

streamlit
pydeck
matplotlib
plotly
httpx

mlflow
optuna

boto3
botocore

requests

shap

pyyaml
python-dotenv